% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{do.call.stash}
\alias{do.call.stash}
\title{Execute a Function Call and Stash the Result}
\usage{
do.call.stash(fun, argslist)
}
\arguments{
\item{fun}{Function to apply. Unlike `do.call`, strings naming the function are not supported}

\item{argslist}{List of arguments to `fun`, in order.}
}
\description{
Similar to `do.call`, `do.call.stash` constructs and executes a function call.
Additionally, `do.call.stash` stores the result indexed by the function body,
its arguments, and its dependencies.
}
\examples{
\dontrun{
heavy_function <- function(x) {
  Sys.sleep(3)
  mean(x)
}

result <- NULL
system.time({result <- do.call.stash(heavy_function, list(c(1:20)))})
result
system.time({result <- do.call.stash(heavy_function, list(c(1:20)))})
result
}

}
