% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{is_stash_pointer}
\alias{is_stash_pointer}
\title{Test if an Object is a stash_pointer}
\usage{
is_stash_pointer(x)
}
\arguments{
\item{x}{object}
}
\value{
Logical
}
\description{
Test if an Object is a stash_pointer
}
