% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{print.stash_pointer}
\alias{print.stash_pointer}
\title{Print Brief Info on a stash_pointer}
\usage{
\method{print}{stash_pointer}(x)
}
\arguments{
\item{x}{}
}
\description{
Print Brief Info on a stash_pointer
}
