% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{print.stash_ref}
\alias{print.stash_ref}
\title{Print Brief Info on a stash_ref}
\usage{
\method{print}{stash_ref}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dot-dot-dot}
}
\description{
Print Brief Info on a stash_ref
}
