% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{stash}
\alias{stash}
\title{Stash an R Object onto Disk}
\usage{
stash(object, dir_path = tempdir(),
  file_name = paste0(sample(c(letters, LETTERS, 0:9), 20, TRUE), collapse
  = ""))
}
\arguments{
\item{object}{an R object; name}

\item{dir_path}{path to a directory;character string}

\item{file_name}{name of the stash file}
}
\value{
a stash_ref obj
}
\description{
Stash an R Object onto Disk
}
